reservation system
project helps to reserve the bus seats for traveling. This system is based on a concept to reserve bus tickets.From this system, the user can view bus list, book tickets and check bus status.While booking tickets, the user has to enter bus number , passenger name , time and departure and arrival place. The system also displays if the seat is already reserved for someone. This project involves various strings and variables.